DD and overloaded blessed globs
Data::Dumper’s Perl implementation was not working with overloaded
blessed globs, which it thought were strings.

The _dump routine was already copying the value to another variable,
but it was using ref \$original or \$original stringified to see
whether it was dealing with a glob.  Since the copy was already there,
it made sense for it to use ref \$copy instead.  So that’s what I
made it do.